Tests/LibWeb/Text/expected/HTML/input-type-change-unchecks-radio-css-checked.txt

zIndex = 1
zIndex = 0
